feat: core performance optimizations (frontend + backend)\n\n- Frontend: image optimization for FeaturedServiceCard (Cloudinary)\n- Frontend: lazy ChatWidget via client-only wrapper\n- Frontend: improved infinite scroll + skeletons on home page\n- Frontend: salons/[id] split to server page + client island\n- Backend: add compression middleware\n- Backend: add Prisma indexes for common queries\n\nDroid-assisted
