+++ backend/prisma/schema.prisma
@@ -68,6 +68,9 @@ model Salon {
   reviews              Review[]
   favorites            Favorite[]
   gallery              GalleryImage[]
+  @@index([approvalStatus])
+  @@index([city])
+  @@index([province])
 }
 
 // New Model for Service Categories
@@ -100,6 +103,8 @@ model Service {
   createdAt      DateTime       @default(now())
   updatedAt      DateTime       @updatedAt
   promotions     Promotion[] // Relation to promotions
+  @@index([salonId])
+  @@index([approvalStatus])
 }
 
 model Review {
@@ -115,6 +120,7 @@ model Review {
   salon          Salon          @relation(fields: [salonId], references: [id])
   bookingId      String         @unique
   booking        Booking        @relation(fields: [bookingId], references: [id])
+  @@index([salonId])
 }
 
 model Booking {
@@ -133,6 +139,9 @@ model Booking {
   service     Service       @relation(fields: [serviceId], references: [id])
   clientPhone String?
   review      Review?
+  @@index([userId])
+  @@index([salonId])
+  @@index([serviceId])
 }
 
 // New Model for Products
@@ -180,6 +189,8 @@ model Conversation {
   updatedAt DateTime  @updatedAt
 
   @@unique([user1Id, user2Id])
+  @@index([user1Id])
+  @@index([user2Id])
 }
 
 model Message {
@@ -193,6 +204,8 @@ model Message {
   conversation   Conversation @relation(fields: [conversationId], references: [id])
   senderId       String
   sender         User         @relation("SentMessages", fields: [senderId], references: [id])
+  @@index([conversationId])
+  @@index([senderId])
 }
 
 model Notification {
@@ -204,6 +217,7 @@ model Notification {
   user      User     @relation(fields: [userId], references: [id])
   bookingId String?
   link      String?
+  @@index([userId])
 }
 
 model ProductOrder {
@@ -222,6 +236,9 @@ model ProductOrder {
   notes         String?
   createdAt     DateTime            @default(now())
   updatedAt     DateTime            @updatedAt
+  @@index([buyerId])
+  @@index([sellerId])
+  @@index([productId])
 }
 
 model Favorite {
@@ -232,6 +249,7 @@ model Favorite {
   salon     Salon    @relation(fields: [salonId], references: [id])
 
   @@id([userId, salonId])
+  @@index([salonId])
 }
 
 model ServiceLike {
@@ -242,6 +260,7 @@ model ServiceLike {
   service   Service  @relation(fields: [serviceId], references: [id])
 
   @@id([userId, serviceId])
+  @@index([serviceId])
 }
 
 model GalleryImage {
@@ -251,6 +270,7 @@ model GalleryImage {
   salonId   String
   salon     Salon    @relation(fields: [salonId], references: [id])
   createdAt DateTime @default(now())
+  @@index([salonId])
 }
 
 enum UserRole {
