+++ frontend/src/app/salons/[id]/SalonProfileClient.tsx
@@ -1,173 +1,121 @@
-// frontend/src/app/salons/[id]/SalonProfileClient.tsx
-'use client';
+"use client";
 
-import { useState, useEffect, useCallback } from 'react';
+import { useEffect, useMemo, useState, useCallback } from 'react';
 import { useRouter } from 'next/navigation';
 import Head from 'next/head';
 import Link from 'next/link';
-import { FaHome, FaArrowLeft, FaHeart, FaWhatsapp, FaGlobe, FaEdit, FaPlus, FaCamera, FaStar } from 'react-icons/fa';
-import { Salon, Service, GalleryImage, Review, Booking } from '@/types';
+import Image from 'next/image';
+import { transformCloudinary } from '@/utils/cloudinary';
+import { FaHome, FaArrowLeft, FaHeart, FaWhatsapp, FaGlobe } from 'react-icons/fa';
+import { Salon, Service, GalleryImage } from '@/types';
 import BookingModal from '@/components/BookingModal';
 import styles from './SalonProfile.module.css';
 import LoadingSpinner from '@/components/LoadingSpinner';
-import { useAuth } from '@/hooks/useAuth';
 import Accordion from '@/components/Accordion';
 import ServiceCard from '@/components/ServiceCard';
 import { toast } from 'react-toastify';
 import { useSocket } from '@/context/SocketContext';
 import ImageLightbox from '@/components/ImageLightbox';
-import EditSalonModal from '@/components/EditSalonModal';
-import ServiceFormModal from '@/components/ServiceFormModal';
-import ReviewModal from '@/components/ReviewModal';
-import GalleryUploadModal from '@/components/GalleryUploadModal';
+import { useAuth } from '@/hooks/useAuth';
+import { useAuthModal } from '@/context/AuthModalContext';
+import { useStartConversation } from '@/hooks/useStartConversation';
 
-interface SalonProfileClientProps {
+type Props = {
   initialSalon: Salon | null;
-  initialServices: Service[];
-  initialGalleryImages: GalleryImage[];
-}
+};
 
-export default function SalonProfileClient({ initialSalon, initialServices, initialGalleryImages }: SalonProfileClientProps) {
+export default function SalonProfileClient({ initialSalon }: Props) {
   const router = useRouter();
   const { authStatus, user } = useAuth();
+  const { openModal } = useAuthModal();
   const socket = useSocket();
+  const { startConversation } = useStartConversation();
 
   const [salon, setSalon] = useState<Salon | null>(initialSalon);
-  const [services, setServices] = useState<Service[]>(initialServices);
-  const [galleryImages, setGalleryImages] = useState<GalleryImage[]>(initialGalleryImages);
+  const [services, setServices] = useState<Service[]>(initialSalon?.services ?? []);
+  const [galleryImages, setGalleryImages] = useState<GalleryImage[]>(initialSalon?.gallery ?? []);
   const [isLoading, setIsLoading] = useState(!initialSalon);
-
   const [selectedService, setSelectedService] = useState<Service | null>(null);
-  const [isLightboxOpen, setLightboxOpen] = useState(false);
-  const [lightboxImageUrl, setLightboxImageUrl] = useState('');
+  const [isLightboxOpen, setIsLightboxOpen] = useState(false);
+  const [lightboxImages, setLightboxImages] = useState<string[]>([]);
+  const [lightboxStartIndex, setLightboxStartIndex] = useState(0);
+  const [currentSlide, setCurrentSlide] = useState(0);
 
-  const [isEditModalOpen, setIsEditModalOpen] = useState(false);
-  const [isServiceModalOpen, setIsServiceModalOpen] = useState(false);
-  const [isReviewModalOpen, setIsReviewModalOpen] = useState(false);
-  const [isGalleryModalOpen, setIsGalleryModalOpen] = useState(false);
-
-  const isOwner = salon && user && (user.id === salon.ownerId || user.role === 'ADMIN');
-
-  const fetchPageData = useCallback(async () => {
-    if (salon?.id) {
-      setIsLoading(true);
-      try {
-        const [salonRes, servicesRes, galleryRes] = await Promise.all([
-          fetch(`/api/salons/${salon.id}`, { credentials: 'include', cache: 'no-store' }),
-          fetch(`/api/services/salon/${salon.id}`),
-          fetch(`/api/gallery/salon/${salon.id}`),
-        ]);
-
-        if (salonRes.ok) setSalon(await salonRes.json());
-        if (servicesRes.ok) setServices(await servicesRes.json());
-        if (galleryRes.ok) setGalleryImages(await galleryRes.json());
-
-      } catch (error) {
-        console.error('Failed to refresh data:', error);
-        toast.error('Could not load latest salon details.');
-      } finally {
-        setIsLoading(false);
-      }
-    }
-  }, [salon?.id]);
+  const heroImagesCount = salon?.heroImages?.length ?? 0;
 
   useEffect(() => {
-    if (socket && salon) {
-      socket.emit('joinSalonRoom', salon.id);
-      socket.on('availabilityUpdate', (data: { isAvailableNow: boolean }) => {
-        setSalon(prev => prev ? { ...prev, isAvailableNow: data.isAvailableNow } : null);
-      });
-      return () => {
-        socket.emit('leaveSalonRoom', salon.id);
-        socket.off('availabilityUpdate');
-      };
-    }
-  }, [socket, salon]);
+    if (heroImagesCount < 2) return;
+    const id = setInterval(() => {
+      setCurrentSlide((prev) => (prev === heroImagesCount - 1 ? 0 : prev + 1));
+    }, 5000);
+    return () => clearInterval(id);
+  }, [heroImagesCount]);
 
-  const openLightbox = (imageUrl: string) => {
-    setLightboxImageUrl(imageUrl);
-    setLightboxOpen(true);
+  useEffect(() => {
+    if (!salon?.id || !socket) return;
+    socket.emit('joinSalonRoom', salon.id);
+    socket.on('availabilityUpdate', (data: { isAvailableNow: boolean }) => {
+      setSalon(prev => prev ? { ...prev, isAvailableNow: data.isAvailableNow } : null);
+    });
+    return () => {
+      socket.emit('leaveSalonRoom', salon.id);
+      socket.off('availabilityUpdate');
+    };
+  }, [socket, salon?.id]);
+
+  const openLightbox = (images: string[], index: number) => {
+    setLightboxImages(images);
+    setLightboxStartIndex(index);
+    setIsLightboxOpen(true);
   };
 
-  const handleBookNowClick = (service: Service) => {
+  const closeLightbox = () => {
+    setIsLightboxOpen(false);
+    setLightboxImages([]);
+    setLightboxStartIndex(0);
+  };
+
+  const handleBookClick = (service: Service) => {
     if (authStatus !== 'authenticated') {
-      router.push('/login');
+      openModal('login');
     } else {
       setSelectedService(service);
     }
   };
 
   const handleSendMessageClick = async () => {
-    if (authStatus !== 'authenticated') {
-      router.push('/login');
+    if (!salon) return;
+    if (user && user.id === salon.ownerId) {
+      toast.error('You cannot message your own salon.');
       return;
     }
-    if (!salon || !user) return;
-    if (user.id === salon.ownerId) {
-      toast.error("You cannot message your own salon.");
-      return;
-    }
-    try {
-      const res = await fetch('/api/chat/conversations', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ recipientId: salon.ownerId }),
-        credentials: 'include',
-      });
-      if (!res.ok) throw new Error('Failed to start conversation.');
-      const conversation = await res.json();
-      router.push(`/chat/${conversation.id}`);
-    } catch (error) {
-      console.error(error);
-      toast.error('Could not start chat. Please try again.');
-    }
+    void startConversation(salon.ownerId, { recipientName: salon.name });
   };
 
   const handleToggleFavorite = async () => {
     if (authStatus !== 'authenticated') {
-      toast.error('You must be logged in to favorite a salon.');
-      router.push('/login');
+      toast.info('Please log in to add to favorites.');
+      openModal('login');
       return;
     }
     if (!salon) return;
+
+    const originalFavoritedState = salon.isFavorited;
+    setSalon(prev => (prev ? { ...prev, isFavorited: !prev.isFavorited } : null));
+
     try {
-      setSalon(prev => (prev ? { ...prev, isFavorited: !prev.isFavorited } : null));
       const res = await fetch(`/api/favorites/toggle/${salon.id}`, {
         method: 'POST',
         credentials: 'include',
       });
-      if (!res.ok) {
-        setSalon(prev => (prev ? { ...prev, isFavorited: !prev.isFavorited } : null)); // Revert on fail
-        throw new Error('Failed to update favorite status.');
-      }
-      toast.success(salon.isFavorited ? 'Removed from favorites.' : 'Added to favorites!');
-    } catch (error) {
+      if (!res.ok) throw new Error('Failed to update favorite status.');
+      const { favorited } = await res.json();
+      toast.success(favorited ? 'Added to favorites!' : 'Removed from favorites.');
+    } catch {
       toast.error('Could not update favorites.');
+      setSalon(prev => (prev ? { ...prev, isFavorited: originalFavoritedState } : null));
     }
   };
-
-  const handleToggleAvailability = async () => {
-    if (!salon) return;
-    try {
-        const res = await fetch(`/api/salons/mine/availability`, {
-            method: 'PATCH',
-            credentials: 'include',
-        });
-        if (!res.ok) throw new Error('Failed to update availability.');
-        const updatedSalon = await res.json();
-        setSalon(s => s ? { ...s, isAvailableNow: updatedSalon.isAvailableNow } : null);
-        toast.success(`Salon is now ${updatedSalon.isAvailableNow ? 'available' : 'unavailable'} for bookings.`);
-    } catch (error) {
-        toast.error('Failed to update availability.');
-    }
-  };
-
-  if (isLoading) return <LoadingSpinner />;
-  if (!salon) return <div style={{ textAlign: 'center', padding: '2rem' }}>Salon not found.</div>;
-
-  const mapSrc = salon.latitude && salon.longitude
-    ? `https://maps.google.com/maps?q=${salon.latitude},${salon.longitude}&hl=es;z=14&output=embed`
-    : '';
 
   const formatBookingType = (type: string) => {
     if (type === 'ONSITE') return 'This salon offers on-site services.';
@@ -175,13 +123,45 @@ export default function SalonProfileClient({ initialSalon, initialServices, init
     if (type === 'BOTH') return 'This salon offers both on-site and mobile services.';
     return 'Not Specified';
   };
-  
-  const operatingDays = salon.operatingHours ? Object.keys(salon.operatingHours) : [];
+
+  const operatingDays = salon?.operatingHours ? Object.keys(salon.operatingHours) : [];
+  const operatingSummary = useMemo(() => {
+    if (!salon?.operatingHours) return '';
+    const entries = Object.entries(salon.operatingHours as Record<string, string>);
+    if (entries.length === 0) return '';
+    const samples = entries.slice(0, 2).map(([day, hours]) => `${day.substring(0,3)} ${hours}`);
+    const extra = entries.length > 2 ? ` +${entries.length - 2} more` : '';
+    return `${samples.join(' • ')}${extra}`;
+  }, [salon?.operatingHours]);
+
+  if (isLoading) return <LoadingSpinner />;
+  if (!salon) return <div style={{textAlign: 'center', padding: '2rem'}}>Salon not found.</div>;
+
+  const heroImages = salon.heroImages || [];
+  const nextSlide = () => setCurrentSlide(prev => (prev === heroImages.length - 1 ? 0 : prev + 1));
+  const prevSlide = () => setCurrentSlide(prev => (prev === 0 ? heroImages.length - 1 : prev - 1));
+
+  const galleryImageUrls = galleryImages.map(img => img.imageUrl);
+  const mapSrc = (() => {
+    if (!salon.latitude || !salon.longitude) return '';
+    const lat = salon.latitude;
+    const lon = salon.longitude;
+    const d = 0.01;
+    const bbox = `${lon - d},${lat - d},${lon + d},${lat + d}`;
+    return `https://www.openstreetmap.org/export/embed.html?bbox=${bbox}&layer=mapnik&marker=${lat},${lon}`;
+  })();
 
   return (
     <>
       <Head>
         <title>{`${salon.name} - The Salon Hub`}</title>
+        <meta name="description" content={`Find the best beauty services at ${salon.name} in ${salon.city}. Book online today!`} />
+        <meta name="keywords" content={`salon, ${salon.name}, ${salon.city}, beauty, haircut, nails, stylist, hairdresser`} />
+        <meta property="og:title" content={`${salon.name} - The Salon Hub`} />
+        <meta property="og:description" content={`Find the best beauty services at ${salon.name} in ${salon.city}. Book online today!`} />
+        <meta property="og:image" content={salon.backgroundImage || '/logo-transparent.png'} />
+        <meta property="og:url" content={`https://thesalonhub.com/salons/${salon.id}`} />
+        <meta name="twitter:card" content="summary_large_image" />
       </Head>
 
       {selectedService && (
@@ -189,61 +169,183 @@ export default function SalonProfileClient({ initialSalon, initialServices, init
           salon={salon}
           service={selectedService}
           onClose={() => setSelectedService(null)}
-          onBookingSuccess={(booking: Booking) => {
+          onBookingSuccess={() => {
             toast.success('Booking request sent! The salon will confirm shortly.');
             setSelectedService(null);
           }}
         />
       )}
-      {isEditModalOpen && salon && (
-        <EditSalonModal
-          salon={salon}
-          onClose={() => setIsEditModalOpen(false)}
-          onSalonUpdate={(updatedSalon) => {
-            setSalon(updatedSalon);
-            setIsEditModalOpen(false);
-          }}
-        />
-      )}
-      {isServiceModalOpen && salon && (
-        <ServiceFormModal
-          salonId={salon.id}
-          onClose={() => setIsServiceModalOpen(false)}
-          onServiceAdded={(newService: Service) => {
-            setServices(prev => [...prev, newService]);
-            setIsServiceModalOpen(false);
-          }}
-        />
-      )}
-      {isReviewModalOpen && salon && (
-        <ReviewModal
-          bookingId="" // Pass a valid booking ID if available
-          onClose={() => setIsReviewModalOpen(false)}
-          onReviewAdded={(newReview: Review) => {
-            setSalon(prev => prev ? { ...prev, reviews: [newReview, ...(prev.reviews || [])] } : null);
-            setIsReviewModalOpen(false);
-          }}
-        />
-      )}
-      {isGalleryModalOpen && salon && (
-        <GalleryUploadModal
-          salonId={salon.id}
-          onClose={() => setIsGalleryModalOpen(false)}
-          onImageAdded={(newImage: GalleryImage) => {
-            setGalleryImages(prev => [...prev, newImage]);
-            // No need to close modal here, user can upload multiple
-          }}
-        />
-      )}
       {isLightboxOpen && (
         <ImageLightbox
-          images={[lightboxImageUrl]}
-          initialImageIndex={0}
-          onClose={() => setLightboxOpen(false)}
+          images={lightboxImages}
+          initialImageIndex={lightboxStartIndex}
+          onClose={closeLightbox}
         />
       )}
 
-      {/* ... The rest of the JSX from page.tsx, adapted to use the state from this component ... */}
+      <div>
+        <div className={styles.stickyHeader}>
+          <div className={styles.navButtonsContainer}>
+            <button onClick={() => router.back()} className={styles.navButton}><FaArrowLeft /> Back</button>
+            <Link href="/" className={styles.navButton}><FaHome /> Home</Link>
+          </div>
+          <h1 className={styles.title}>{salon.name}</h1>
+          <div className={styles.headerSpacer}>
+            {authStatus === 'authenticated' && (
+              <button onClick={handleToggleFavorite} className={`${styles.favoriteButton} ${salon.isFavorited ? styles.favorited : ''}`}>
+                <FaHeart />
+              </button>
+            )}
+          </div>
+        </div>
+
+        <div className={styles.container}>
+          <div className={styles.heroSlider}>
+            {heroImages.length > 0 ? (
+              <>
+                <div className={styles.sliderWrapper} style={{ transform: `translateX(-${currentSlide * 100}%)` }}>
+                  {heroImages.map((img: string, index: number) => (
+                    <div key={index} className={styles.slide}>
+                      <Image
+                        src={transformCloudinary(img, { width: 1200, quality: 'auto', format: 'auto', crop: 'fill' })}
+                        alt={`${salon.name} gallery image ${index + 1}`}
+                        className={styles.heroImage}
+                        fill
+                        sizes="(max-width: 1024px) 100vw, 1024px"
+                        priority={index === 0}
+                      />
+                    </div>
+                  ))}
+                </div>
+                {heroImages.length > 1 && (
+                  <>
+                    <button onClick={prevSlide} className={`${styles.sliderButton} ${styles.prev}`}>❮</button>
+                    <button onClick={nextSlide} className={`${styles.sliderButton} ${styles.next}`}>❯</button>
+                  </>
+                )}
+              </>
+            ) : (
+              <div className={styles.placeholderHero}>
+                <p>No images available for this salon</p>
+              </div>
+            )}
+            <div className={styles.heroOverlay}>
+              <p className={styles.heroLocation}>{salon.town}, {salon.province}</p>
+              {operatingSummary && (
+                <p className={styles.heroHours}>Hours: {operatingSummary}</p>
+              )}
+              {!!salon.isAvailableNow && (
+                <div className={styles.availabilityIndicator}>
+                  <span className={styles.availabilityDot}></span>
+                  Available Now
+                </div>
+              )}
+            </div>
+          </div>
+
+          <div className={styles.profileLayout}>
+            <div className={styles.mainContent}>
+              {galleryImages.length > 0 && (
+                <section id="gallery-section">
+                  <h2 className={styles.sectionTitle}>Gallery</h2>
+                  <div className={styles.galleryGrid}>
+                    {galleryImages.map((image, index) => (
+                      <div key={image.id} className={styles.galleryItem} onClick={() => openLightbox(galleryImageUrls, index)}>
+                        <Image
+                          src={transformCloudinary(image.imageUrl, { width: 400, quality: 'auto', format: 'auto', crop: 'fill' })}
+                          alt={image.caption || 'Salon work'}
+                          className={styles.galleryImage}
+                          fill
+                          sizes="(max-width: 768px) 50vw, 200px"
+                        />
+                      </div>
+                    ))}
+                  </div>
+                </section>
+              )}
+              <section id="services-section">
+                <h2 className={styles.sectionTitle}>Services</h2>
+                <div className={styles.servicesGrid}>
+                  {services.map((service) => (
+                    <ServiceCard 
+                      key={service.id} 
+                      service={service} 
+                      onBook={handleBookClick}
+                      onSendMessage={handleSendMessageClick}
+                      onImageClick={openLightbox}
+                    />
+                  ))}
+                </div>
+              </section>
+
+              <section>
+                <h2 className={styles.sectionTitle}>Details</h2>
+
+                <Accordion title="Operating Hours">
+                  {salon.operatingHours ? (
+                    <ul>
+                      {operatingDays.map(day => (
+                        <li key={day}>
+                          <span>{day.charAt(0).toUpperCase() + day.slice(1)}</span>
+                          <strong>{salon.operatingHours![day]}</strong>
+                        </li>
+                      ))}
+                    </ul>
+                  ) : <p>Operating hours not listed.</p>}
+                </Accordion>
+                <Accordion title="Service Type">
+                  <p>{formatBookingType(salon.bookingType)}</p>
+                  {salon.offersMobile && salon.mobileFee && (
+                    <p style={{marginTop: '0.5rem'}}>Mobile service fee: <strong>R{salon.mobileFee.toFixed(2)}</strong></p>
+                  )}
+                </Accordion>
+                <Accordion title="Location & Contact">
+                  <p><strong>Address:</strong> {salon.address || `${salon.town}, ${salon.city}, ${salon.province}`}</p>
+                  {authStatus === 'authenticated' ? (
+                    <>
+                      {salon.contactEmail && <p><strong>Email:</strong> <a href={`mailto:${salon.contactEmail}`}>{salon.contactEmail}</a></p>}
+                      {salon.phoneNumber && <p><strong>Phone:</strong> <a href={`tel:${salon.phoneNumber}`}>{salon.phoneNumber}</a></p>}
+                      {salon.whatsapp && <p><strong><FaWhatsapp /> WhatsApp:</strong> <a href={`https://wa.me/${salon.whatsapp}`} target="_blank" rel="noopener noreferrer">{salon.whatsapp}</a></p>}
+                      {salon.website && <p><strong><FaGlobe /> Website:</strong> <a href={salon.website} target="_blank" rel="noopener noreferrer">{salon.website}</a></p>}
+                      {mapSrc && (
+                        <div className={styles.mapContainer}>
+                          <iframe
+                            width="100%"
+                            height="300"
+                            style={{ border: 0, borderRadius: '0.5rem', marginTop: '1rem' }}
+                            loading="lazy"
+                            allowFullScreen
+                            src={mapSrc}>
+                          </iframe>
+                        </div>
+                      )}
+                    </>
+                  ) : (
+                    <p className={styles.loginPrompt}>
+                      <Link href="/login">Log in</Link> to view detailed contact information and map.
+                    </p>
+                  )}
+                </Accordion>
+                <Accordion title={`Reviews (${salon.reviews?.length || 0})`}>
+                  {salon.reviews && salon.reviews.length > 0 ? (
+                    <div>
+                      {salon.reviews.map(review => (
+                        <div key={review.id} style={{borderBottom: '1px dotted var(--color-border)', paddingBottom: '1rem', marginBottom: '1rem'}}>
+                          <div style={{display: 'flex', justifyContent: 'space-between', alignItems: 'center'}}>
+                            <strong>{review.author.firstName} {review.author.lastName.charAt(0)}.</strong>
+                            <span style={{color: 'var(--accent-gold)'}}>{'★'.repeat(review.rating)}{'☆'.repeat(5 - review.rating)}</span>
+                          </div>
+                          <p style={{fontStyle: 'italic', marginTop: '0.5rem'}}>&quot;{review.comment}&quot;</p>
+                        </div>
+                      ))}
+                    </div>
+                  ) : <p>No reviews yet.</p>}
+                </Accordion>
+              </section>
+            </div>
+          </div>
+        </div>
+      </div>
     </>
   );
 }
