+++ frontend/src/components/FeaturedServiceCard.tsx
@@ -4,6 +4,7 @@ import Link from 'next/link';
 import Image from 'next/image';
 import { Service } from '@/types';
 import styles from './FeaturedServiceCard.module.css';
+import { transformCloudinary } from '@/utils/cloudinary';
 
 interface FeaturedServiceCardProps {
   service: Service & { salon: { id: string; name: string; city: string; province: string } };
@@ -14,7 +15,7 @@ export default function FeaturedServiceCard({ service }: FeaturedServiceCardProp
     <Link href={`/salons/${service.salon.id}`} className={styles.card}>
       <div className={styles.cardImageWrapper}>
         <Image
-          src={service.images[0] || 'https://via.placeholder.com/300x150'}
+          src={transformCloudinary(service.images[0] || 'https://via.placeholder.com/300x150', { width: 600, quality: 'auto', format: 'auto', crop: 'fill' })}
           alt={service.title}
           className={styles.cardImage}
           fill
